+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Upcoming version
 
 ### Added
+
+- Added `available()` API in `AddressAllocator` to allow
+  getting the available memories after `allocate/free()`s
+
 ### Changed
 ### Fixed
 ### Removed